Remove twitter-widgets-loader and update widgets.js
The widgets loader just ensures that the widgets script is only loaded once, but this is redundant since htmlDependency() will also take care of that for us.

This also fixes #5

The new widgets.js has new formatting and uses an <iframe> instead of the previous approach, so I also had to update the screen shot javascript.

Additionally, I extracted the screenshot page for easier testing.